# gin-vue-admin/server/docs/swagger.yaml
definitions:
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  starNote.NoteCategory:
    properties:
      ID:
        description: 主键ID
        type: integer
      color:
        description: 颜色
        type: string
      createdAt:
        description: 创建时间
        type: string
      icon:
        description: 图标
        type: string
      name:
        description: 名称
        type: string
      updatedAt:
        description: 更新时间
        type: string
    required:
    - name
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.9.0
paths:
...
  /nc/createNoteCategory:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建记事本分类管理
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/starNote.NoteCategory'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建记事本分类管理
      tags:
      - NoteCategory
  /nc/deleteNoteCategory:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除记事本分类管理
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/starNote.NoteCategory'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除记事本分类管理
      tags:
      - NoteCategory
  /nc/deleteNoteCategoryByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除记事本分类管理
      tags:
      - NoteCategory
  /nc/findNoteCategory:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询记事本分类管理
        in: query
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/starNote.NoteCategory'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询记事本分类管理
      tags:
      - NoteCategory
  /nc/getNoteCategoryList:
    get:
      consumes:
      - application/json
      parameters:
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: createdAtRange
        type: array
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取记事本分类管理列表
      tags:
      - NoteCategory
...
  /nc/updateNoteCategory:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新记事本分类管理
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/starNote.NoteCategory'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新记事本分类管理
      tags:
      - NoteCategory
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...